plag3/requirements.txt
Flask==2.2.5
PyPDF2==3.0.1
python-docx==0.8.11
werkzeug==2.3.7
